Refactor: 클래스별 책임 분산
