+++ src/main/java/controller/LottoController.java
@@ -1,6 +1,6 @@
 package controller;
 
-import domain.Lotto;
+import domain.*;
 import dto.LottoDto;
 import service.LottoService;
 import view.*;
@@ -19,7 +19,7 @@ public class LottoController {
         int lottoAmount = lottoService.calculateGetLottoAmount(purchaseAmount);
         lottoInputView.printEmptyLine();
 
-        List<Lotto> lottoList = createLottoList(lottoAmount);
+        List<Lotto> lottoList = lottoService.createLottoList(lottoAmount);
         lottoOutputView.printLottoAmount(lottoAmount);
 
         for (Lotto lotto : lottoList) {
@@ -27,22 +27,8 @@ public class LottoController {
         }
     }
 
-    private List<Lotto> createLottoList(int lottoAmount) {
-        List<Lotto> lottoList = new ArrayList<>();
-
-        for (int i = 0; i < lottoAmount; i++) {
-            lottoList.add(getLotto());
-        }
-
-        return List.copyOf(lottoList);
-    }
-
     private int getPurchaseAmount() {
         return lottoInputView.getPurchaseAmount();
-    }
-
-    private Lotto getLotto() {
-        return Lotto.getRandomLotto();
     }
 
 }
