+++ src/main/java/domain/Lotto.java
@@ -3,8 +3,6 @@ package domain;
 import java.util.*;
 
 public class Lotto {
-
-    private final static int LOTTO_SIZE = 6;
     private final List<LottoNumber> lottoNumbers;
 
     public Lotto(List<LottoNumber> numbers) {
@@ -13,25 +11,6 @@ public class Lotto {
 
     public List<LottoNumber> getLottoNumbers() {
         return List.copyOf(lottoNumbers);
-    }
-
-    public static Lotto getRandomLotto() {
-        List<LottoNumber> numbers = new ArrayList<>();
-
-        while (numbers.size() < LOTTO_SIZE) {
-            LottoNumber randomNumber = LottoNumber.getRandomLottoNumber();
-            putIfAbsent(numbers, randomNumber);
-        }
-
-        Collections.sort(numbers, Comparator.comparing(LottoNumber::getValue));
-
-        return new Lotto(numbers);
-    }
-
-    private static void putIfAbsent(List<LottoNumber> numbers, LottoNumber randomNumber) {
-        if (!numbers.contains(randomNumber)) {
-            numbers.add(randomNumber);
-        }
     }
 
 }
