+++ src/main/java/service/LottoService.java
@@ -1,5 +1,9 @@
 package service;
 
+import domain.*;
+import java.util.*;
+import java.util.stream.*;
+
 public class LottoService {
 
     private final static int LOTTO_PRICE = 1_000;
@@ -16,6 +20,12 @@ public class LottoService {
         if (purchaseAmount % LOTTO_PRICE != 0) {
             throw new IllegalArgumentException("구입 금액 단위는 " + LOTTO_PRICE + "입니다.");
         }
+    }
+
+    public List<Lotto> createLottoList(int lottoAmount) {
+        return IntStream.range(0, lottoAmount)
+                .mapToObj(i -> LottoMachine.getRandomLotto())
+                .collect(Collectors.toList());
     }
 
 }
